refactor(banner): convert to fixed position and improve stories
- Convert Banner to fixed position component with proper z-index
- Migrate Banner stories from TS to TSX for better readability
- Add MdAnnouncement icon to Banner stories
- Improve Banner stories structure and styling
